third_party/WebKit/LayoutTests/editing/selection/node-removal-2.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
  assert_selection(
    '^foo<span>b|ar</span>baz',
    selection => {
      const span = selection.document.querySelector('span');
      span.parentNode.removeChild(span);
      selection.document.execCommand('selectAll');
    },
    '^foobaz|');
}, 'selectAll after node remove');
</script>
